Don't send buffers to unknown clients
Whenever autoadd is not enabled and an unknown client connects to ZNC,
ZNC will proceed to indiscriminately send the contents of all buffers
to the client. However, this behaviour keeps repeating each time the
clients connects again: as this module is expected to be used with
AutoClearChanBuffer and AutoClearQueryBuffer both off, the buffers
aren't cleared, and ClientBuffer does not prevent ZNC from re-sending
the buffers on each client connect.

ClientBuffer should instead prevent ZNC from sending buffer contents
to unidentified clients.

Fixes #17.